requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.77.0
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
